+++ src/pages/ProductPage/components/ProductDetail.scss
